updated scan tools
